mcp-server/pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "local-latent-mcp"
version = "0.1.0"
description = "MCP v1 server for Local Latent Containers"
requires-python = ">=3.11"
authors = [{ name = "Silent Architect" }]
dependencies = [
  "fastapi>=0.115.5",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "httpx>=0.27",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "qdrant-client>=1.9",
  "minio>=7.2",
  "pyyaml>=6.0",
  "orjson>=3.10",
  "python-json-logger>=2.0",
  "tenacity>=8.3",
  "prometheus-client>=0.20",
  "alembic>=1.13"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "ruff>=0.6",
  "mypy>=1.10",
  "psycopg[binary]>=3.2",
  "structlog>=24.1"
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_return_any = true
